[REFACTORING] Frontend
