+++ frontend/src/pages/components/AdminNavbar.js
@@ -7,8 +7,8 @@ import {
 } from "@heroicons/react/24/outline";
 import { useEffect, useState } from "react";
 import { useLocation } from "react-router-dom";
-import { useAuthContext } from "../../hooks/useAuthContext";
-import useGetUser from "../../hooks/useGetUser";
+import { useAuthContext } from "hooks/useAuthContext";
+import useGetUser from "hooks/useGetUser";
 import AdminNavLinks from "./AdminNavLinks";
 
 const AdminNavbar = () => {
