+++ frontend/src/pages/components/FollowButton.js
@@ -1,6 +1,6 @@
 import { ExclamationTriangleIcon } from "@heroicons/react/24/outline";
 import { useState } from "react";
-import useFollowOrUnfolllow from "../../hooks/useFollowOrUnfollow";
+import useFollowOrUnfolllow from "hooks/useFollowOrUnfollow";
 
 const FollowButton = ({ followingState, id }) => {
   const [isFollow, setIsfolllow] = useState(followingState);
