+++ frontend/src/pages/lesson/components/HeaderSection.js
@@ -1,7 +1,7 @@
 import { DocumentTextIcon } from "@heroicons/react/24/outline";
 import { ArrowLeftIcon } from "@heroicons/react/24/solid";
 import { Link } from "react-router-dom";
-import { useAuthContext } from "../../../hooks/useAuthContext";
+import { useAuthContext } from "hooks/useAuthContext";
 
 const HeaderSection = ({ current, all, title }) => {
   const { user } = useAuthContext();
