+++ frontend/src/pages/signup/Signup.js
@@ -1,10 +1,10 @@
 import { useForm } from "react-hook-form";
 import { Link, useNavigate } from "react-router-dom";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { authInputSchema } from "../../validations/authInputValidation";
-import learningSvg from "../../images/learning_sketching_nd4f.svg";
+import { authInputSchema } from "validations/authInputValidation";
+import learningSvg from "images/learning_sketching_nd4f.svg";
 import SignupForm from "./components/SingupForm";
-import useSignup from "../../hooks/useSignup";
+import useSignup from "hooks/useSignup";
 
 const Signup = () => {
   const { signup, error, isLoading } = useSignup();
